% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_summary_table.R
\name{make_summary_tbl}
\alias{make_summary_tbl}
\title{Create Summary Table for Trial Data}
\usage{
make_summary_tbl(d)
}
\arguments{
\item{d}{Data frame. Trial dataset with an 'arm' column.}
}
\value{
A gtsummary table object with descriptive statistics by arm.
}
\description{
Generates a formatted summary table using gtsummary, stratified by treatment arm.
}
